+++ src/durationPerHour/index.js
@@ -2,10 +2,10 @@ import isBefore from '../isBefore/index.js'
 import toDate from '../toDate/index.js'
 
 /**
- * @name durationsPerHour
+ * @name durationPerHour
  * @category Hour Helpers
- * @summary Get the durations a time range lies in each full hour
- * @description Get the durations a time range lies in each full hour
+ * @summary Allocate the duration of a time range to all hours entered into
+ * @description Allocate the duration of a time range to all hours entered into
  *
  * @param {Date | String | Number} start - the start of the time range
  * @param {Date | String | Number} end - the end of the time range
@@ -16,7 +16,7 @@ import toDate from '../toDate/index.js'
  * // How much of the time between 07:33h and 10:43h falls into each full hour?
  * const start = new Date('2017-11-15T07:33:00.000+0000')
  * const end = new Date('2017-11-15T10:43:00.000+0000')
- * durationsPerHour(start, end)
+ * durationPerHour(start, end)
  * // => {
  * //   1510729200000: 1620000,
  * //   1510732800000: 3600000,
@@ -44,7 +44,7 @@ export default function (dirtyStart, dirtyEnd) {
     return { [startOfFirstHour]: end - start }
   }
 
-  const durationsPerHour = {
+  const durationPerHour = {
     [startOfFirstHour]: startOfFirstHour + hour - start,
     [startOfLastHour]: end - startOfLastHour
   }
@@ -52,5 +52,5 @@ export default function (dirtyStart, dirtyEnd) {
   const amountOfCompleteHours = (startOfLastHour - startOfFirstHour) / hour - 1
   const keys = [...Array(amountOfCompleteHours)].map((_, key) => startOfFirstHour + (key + 1) * hour)
 
-  return keys.reduce((acc, key) => Object.assign(acc, { [key]: hour }), durationsPerHour)
+  return keys.reduce((acc, key) => Object.assign(acc, { [key]: hour }), durationPerHour)
 }